use dataloader for eval hooks
