+++ mmdet/apis/__init__.py
@@ -1,9 +1,10 @@
 from .inference import (async_inference_detector, inference_detector,
                         init_detector, show_result, show_result_pyplot)
+from .test import multi_gpu_test, single_gpu_test
 from .train import get_root_logger, set_random_seed, train_detector
 
 __all__ = [
     'get_root_logger', 'set_random_seed', 'train_detector', 'init_detector',
     'async_inference_detector', 'inference_detector', 'show_result',
-    'show_result_pyplot'
+    'show_result_pyplot', 'multi_gpu_test', 'single_gpu_test'
 ]
