+++ mmdet/core/evaluation/eval_hooks.py
@@ -1,72 +1,50 @@
-import os
 import os.path as osp
 
-import mmcv
-import torch
-import torch.distributed as dist
-from mmcv.parallel import collate, scatter
 from mmcv.runner import Hook
-from torch.utils.data import Dataset
+from torch.utils.data import DataLoader
 
 
 class DistEvalHook(Hook):
+    """Distributed evaluation hook.
 
-    def __init__(self, dataset, interval=1, **eval_kwargs):
-        from mmdet import datasets
-        if isinstance(dataset, Dataset):
-            self.dataset = dataset
-        elif isinstance(dataset, dict):
-            self.dataset = datasets.build_dataset(dataset, {'test_mode': True})
-        else:
+    Attributes:
+        dataloader (DataLoader): A PyTorch dataloader.
+        interval (int): Evaluation interval (by epochs). Default: 1.
+        tmpdir (str | None): Temporary directory to save the results of all
+            processes. Default: None.
+        gpu_collect (bool): Whether to use gpu or cpu to collect results.
+            Default: False.
+    """
+
+    def __init__(self,
+                 dataloader,
+                 interval=1,
+                 gpu_collect=False,
+                 **eval_kwargs):
+        if not isinstance(dataloader, DataLoader):
             raise TypeError(
-                'dataset must be a Dataset object or a dict, not {}'.format(
-                    type(dataset)))
+                'dataloader must be a pytorch DataLoader, but got {}'.format(
+                    type(dataloader)))
+        self.dataloader = dataloader
         self.interval = interval
+        self.gpu_collect = gpu_collect
         self.eval_kwargs = eval_kwargs
 
     def after_train_epoch(self, runner):
         if not self.every_n_epochs(runner, self.interval):
             return
-        runner.model.eval()
-        results = [None for _ in range(len(self.dataset))]
-        if runner.rank == 0:
-            prog_bar = mmcv.ProgressBar(len(self.dataset))
-        for idx in range(runner.rank, len(self.dataset), runner.world_size):
-            data = self.dataset[idx]
-            data_gpu = scatter(
-                collate([data], samples_per_gpu=1),
-                [torch.cuda.current_device()])[0]
-
-            # compute output
-            with torch.no_grad():
-                result = runner.model(
-                    return_loss=False, rescale=True, **data_gpu)
-            results[idx] = result
-
-            batch_size = runner.world_size
-            if runner.rank == 0:
-                for _ in range(batch_size):
-                    prog_bar.update()
-
+        from mmdet.apis import multi_gpu_test
+        results = multi_gpu_test(
+            runner.model,
+            self.dataloader,
+            tmpdir=osp.join(runner.work_dir, '.eval_hook'),
+            gpu_collect=self.gpu_collect)
         if runner.rank == 0:
             print('\n')
-            dist.barrier()
-            for i in range(1, runner.world_size):
-                tmp_file = osp.join(runner.work_dir, 'temp_{}.pkl'.format(i))
-                tmp_results = mmcv.load(tmp_file)
-                for idx in range(i, len(results), runner.world_size):
-                    results[idx] = tmp_results[idx]
-                os.remove(tmp_file)
             self.evaluate(runner, results)
-        else:
-            tmp_file = osp.join(runner.work_dir,
-                                'temp_{}.pkl'.format(runner.rank))
-            mmcv.dump(results, tmp_file)
-            dist.barrier()
-        dist.barrier()
 
     def evaluate(self, runner, results):
-        eval_res = self.dataset.evaluate(
+        eval_res = self.dataloader.dataset.evaluate(
             results, logger=runner.logger, **self.eval_kwargs)
         for name, val in eval_res.items():
             runner.log_buffer.output[name] = val
